+++ R/utils-assertions.R
@@ -1,19 +1,25 @@
-.pipes <- c("%>%")
+.pipes <- c("%>%", "|>")
 
 is_pipe <- function(x) {
   if (is.character(x)) x <- parse(text = x)[[1]]
   ifelse(is.call(x), as.character(x[[1]]) %in% .pipes, FALSE)
 }
 
-.infixes_assign <- c("<-", "<<-", "->", "->>", "=")
+# From `?base::Syntax` but lightly organized into groups
+.infixes_assign <- c("<-", "<<-", "->", "->>", "=", ":=")
 .infixes_comp <- c("==", "!=", ">", ">=", "<", "<=")
+.infixes_operator <- c(
+  "$", "@", "[", "[[", "^", "-", "+", ":", "*", "/",
+  "!", "&", "&&", "|", "||", "~", "?"
+)
 .infixes <- c(
-  "+", "-", "*", "/", "^", "$", "[", "[[", "!", "%%", "%/%", "%in%",
+  "%%", "%/%", "%in%",
   .infixes_assign,
-  .infixes_comp
+  .infixes_comp,
+  .infixes_operator
 )
 
-is_infix <- function(x, infix_vals = .infixes) {
+is_infix <- function(x, infix_vals = .infixes, exact = FALSE) {
 
   tryCatch({
     if (is.character(x)) {
@@ -26,7 +32,12 @@ is_infix <- function(x, infix_vals = .infixes) {
       return(FALSE)
     }
 
-    any(as.character(out[[1]]) %in% infix_vals)
+    call_text <- as.character(out[[1]])
+
+    is_known_infix <- any(call_text %in% infix_vals)
+    if (exact) return(is_known_infix)
+
+    is_known_infix || any(grepl("^%.*%$", call_text))
   }, error = function(e) {
     # x is not an infix
     FALSE
@@ -34,5 +45,5 @@ is_infix <- function(x, infix_vals = .infixes) {
 }
 
 is_infix_assign <- function(x) {
-  is_infix(x, infix_vals = .infixes_assign)
+  is_infix(x, infix_vals = .infixes_assign, exact = TRUE)
 }